+++ src/OperatorModdedOperations/MelonAssemblyInfo.cs
@@ -4,7 +4,7 @@ using MelonLoader;
 [assembly: MelonInfo(
     typeof(CerberusNativeTabFix),
     "OPERATOR: Modded Operations",
-    "0.3.30",
+    "0.3.31",
     "OPERATOR Modding Project")]
 [assembly: MelonProcess("OPERATOR")]
 [assembly: MelonPlatform(MelonPlatformAttribute.CompatiblePlatforms.WINDOWS_X64)]